requirements.txt
aio-pika==9.5.5
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiormq==6.8.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
black==25.1.0
ccxt==4.4.86
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.3
exceptiongroup==1.3.0
fast-depends==2.4.12
faststream==0.5.42
frozenlist==1.6.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.6
packaging==25.0
pamqp==3.3.0
pathspec==0.12.1
pika==1.3.2
platformdirs==4.3.8
propcache==0.3.1
pycares==4.8.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
ruff==0.11.12
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
watchfiles==1.0.5
yarl==1.20.0
